+++ src/flash-nrf52.cpp
@@ -1,5 +1,5 @@
 /**
- * @file flash.cpp
+ * @file flash-nrf52.cpp
  * @author Bernd Giesecke (bernd.giesecke@rakwireless.com)
  * @brief Initialize, read and write parameters from/to internal flash memory
  * @version 0.1
@@ -23,15 +23,13 @@ const char settings_name[] = "RAK";
 
 File lora_file(InternalFS);
 
-void flash_reset(void);
-
-bool init_flash_done = false;
+void flash_int_reset(void);
 
 /**
  * @brief Initialize access to nRF52 internal file system
  * 
  */
-void init_flash(void)
+void init_int_flash(void)
 {
 	if (init_flash_done)
 	{
@@ -115,7 +113,7 @@ void init_flash(void)
  * @return boolean 
  * 			result of saving
  */
-boolean save_settings(void)
+boolean save_int_settings(void)
 {
 	bool result = true;
 	// Read saved content
@@ -155,7 +153,7 @@ boolean save_settings(void)
  * @brief Reset content of the filesystem
  * 
  */
-void flash_reset(void)
+void flash_int_reset(void)
 {
 	InternalFS.format();
 	if (lora_file.open(settings_name, FILE_O_WRITE))
@@ -165,74 +163,6 @@ void flash_reset(void)
 		lora_file.flush();
 		lora_file.close();
 	}
-}
-
-/**
- * @brief Printout of all settings
- * 
- */
-void log_settings(void)
-{
-	API_LOG("FLASH", "Saved settings:");
-	API_LOG("FLASH", "000 Marks: %02X %02X", g_lorawan_settings.valid_mark_1, g_lorawan_settings.valid_mark_2);
-	API_LOG("FLASH", "002 Dev EUI %02X%02X%02X%02X%02X%02X%02X%02X", g_lorawan_settings.node_device_eui[0], g_lorawan_settings.node_device_eui[1],
-			g_lorawan_settings.node_device_eui[2], g_lorawan_settings.node_device_eui[3],
-			g_lorawan_settings.node_device_eui[4], g_lorawan_settings.node_device_eui[5],
-			g_lorawan_settings.node_device_eui[6], g_lorawan_settings.node_device_eui[7]);
-	API_LOG("FLASH", "010 App EUI %02X%02X%02X%02X%02X%02X%02X%02X", g_lorawan_settings.node_app_eui[0], g_lorawan_settings.node_app_eui[1],
-			g_lorawan_settings.node_app_eui[2], g_lorawan_settings.node_app_eui[3],
-			g_lorawan_settings.node_app_eui[4], g_lorawan_settings.node_app_eui[5],
-			g_lorawan_settings.node_app_eui[6], g_lorawan_settings.node_app_eui[7]);
-	API_LOG("FLASH", "018 App Key %02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
-			g_lorawan_settings.node_app_key[0], g_lorawan_settings.node_app_key[1],
-			g_lorawan_settings.node_app_key[2], g_lorawan_settings.node_app_key[3],
-			g_lorawan_settings.node_app_key[4], g_lorawan_settings.node_app_key[5],
-			g_lorawan_settings.node_app_key[6], g_lorawan_settings.node_app_key[7],
-			g_lorawan_settings.node_app_key[8], g_lorawan_settings.node_app_key[9],
-			g_lorawan_settings.node_app_key[10], g_lorawan_settings.node_app_key[11],
-			g_lorawan_settings.node_app_key[12], g_lorawan_settings.node_app_key[13],
-			g_lorawan_settings.node_app_key[14], g_lorawan_settings.node_app_key[15]);
-	API_LOG("FLASH", "034 Dev Addr %08lX", g_lorawan_settings.node_dev_addr);
-	API_LOG("FLASH", "038 NWS Key %02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
-			g_lorawan_settings.node_nws_key[0], g_lorawan_settings.node_nws_key[1],
-			g_lorawan_settings.node_nws_key[2], g_lorawan_settings.node_nws_key[3],
-			g_lorawan_settings.node_nws_key[4], g_lorawan_settings.node_nws_key[5],
-			g_lorawan_settings.node_nws_key[6], g_lorawan_settings.node_nws_key[7],
-			g_lorawan_settings.node_nws_key[8], g_lorawan_settings.node_nws_key[9],
-			g_lorawan_settings.node_nws_key[10], g_lorawan_settings.node_nws_key[11],
-			g_lorawan_settings.node_nws_key[12], g_lorawan_settings.node_nws_key[13],
-			g_lorawan_settings.node_nws_key[14], g_lorawan_settings.node_nws_key[15]);
-	API_LOG("FLASH", "054 Apps Key %02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
-			g_lorawan_settings.node_apps_key[0], g_lorawan_settings.node_apps_key[1],
-			g_lorawan_settings.node_apps_key[2], g_lorawan_settings.node_apps_key[3],
-			g_lorawan_settings.node_apps_key[4], g_lorawan_settings.node_apps_key[5],
-			g_lorawan_settings.node_apps_key[6], g_lorawan_settings.node_apps_key[7],
-			g_lorawan_settings.node_apps_key[8], g_lorawan_settings.node_apps_key[9],
-			g_lorawan_settings.node_apps_key[10], g_lorawan_settings.node_apps_key[11],
-			g_lorawan_settings.node_apps_key[12], g_lorawan_settings.node_apps_key[13],
-			g_lorawan_settings.node_apps_key[14], g_lorawan_settings.node_apps_key[15]);
-	API_LOG("FLASH", "070 OTAA %s", g_lorawan_settings.otaa_enabled ? "enabled" : "disabled");
-	API_LOG("FLASH", "071 ADR %s", g_lorawan_settings.adr_enabled ? "enabled" : "disabled");
-	API_LOG("FLASH", "072 %s Network", g_lorawan_settings.public_network ? "Public" : "Private");
-	API_LOG("FLASH", "073 Dutycycle %s", g_lorawan_settings.duty_cycle_enabled ? "enabled" : "disabled");
-	API_LOG("FLASH", "074 Repeat time %ld", g_lorawan_settings.send_repeat_time);
-	API_LOG("FLASH", "078 Join trials %d", g_lorawan_settings.join_trials);
-	API_LOG("FLASH", "079 TX Power %d", g_lorawan_settings.tx_power);
-	API_LOG("FLASH", "080 DR %d", g_lorawan_settings.data_rate);
-	API_LOG("FLASH", "081 Class %d", g_lorawan_settings.lora_class);
-	API_LOG("FLASH", "082 Subband %d", g_lorawan_settings.subband_channels);
-	API_LOG("FLASH", "083 Auto join %s", g_lorawan_settings.auto_join ? "enabled" : "disabled");
-	API_LOG("FLASH", "084 Fport %d", g_lorawan_settings.app_port);
-	API_LOG("FLASH", "085 %s Message", g_lorawan_settings.confirmed_msg_enabled ? "Confirmed" : "Unconfirmed");
-	API_LOG("FLASH", "086 Region %s", region_names[g_lorawan_settings.lora_region]);
-	API_LOG("FLASH", "087 Mode %s", g_lorawan_settings.lorawan_enable ? "LPWAN" : "P2P");
-	API_LOG("FLASH", "088 P2P frequency %ld", g_lorawan_settings.p2p_frequency);
-	API_LOG("FLASH", "092 P2P TX Power %d", g_lorawan_settings.p2p_tx_power);
-	API_LOG("FLASH", "093 P2P BW %d", g_lorawan_settings.p2p_bandwidth);
-	API_LOG("FLASH", "094 P2P SF %d", g_lorawan_settings.p2p_sf);
-	API_LOG("FLASH", "095 P2P CR %d", g_lorawan_settings.p2p_cr);
-	API_LOG("FLASH", "096 P2P Preamble length %d", g_lorawan_settings.p2p_preamble_len);
-	API_LOG("FLASH", "097 P2P Symbol Timeout %d", g_lorawan_settings.p2p_symbol_timeout);
 }
 
 /**